+++ tests/nn/beam_search_test.py
@@ -119,6 +119,42 @@ class BeamSearchTest(AllenNlpTestCase):
         for key, array in expected_finished_state.items():
             np.testing.assert_allclose(state[key].numpy(), array)
 
+    def test_diff_shape_state(self):
+        state = {}
+        state["decoder_hidden"] = torch.tensor(
+            [[1, 0, 1], [2, 0, 1], [0, 0, 1], [1, 1, 1], [0, 0, 0]]
+        )
+        state["decoder_hidden"] = state["decoder_hidden"].unsqueeze(0).repeat(2, 1, 1)
+        # shape: (2, batch_size, 3)
+
+        seq = [
+            [1, 0, 1],
+            [1, 0, 1],
+            [1, 0, 1],
+            [2, 0, 1],
+            [2, 0, 1],
+            [2, 0, 1],
+            [0, 0, 1],
+            [0, 0, 1],
+            [0, 0, 1],
+            [1, 1, 1],
+            [1, 1, 1],
+            [1, 1, 1],
+            [0, 0, 0],
+            [0, 0, 0],
+            [0, 0, 0],
+        ]
+        seq = [seq] * 2
+        expected_finished_state = {}
+        expected_finished_state["decoder_hidden"] = np.array(seq)
+        # shape: (2, batch_size x beam_size, 3)
+
+        self._check_results(state=state)
+
+        # check finished state.
+        for key, array in expected_finished_state.items():
+            np.testing.assert_allclose(state[key].numpy(), array)
+
     def test_batch_size_of_one(self):
         self._check_results(batch_size=1)